bcftools mpileup: align probaln_glocal FP width with htslib
Close the documented single-ULP probaln_glocal rounding drift that
perturbed INFO/I16 and QS values on a handful of deep indel rows at long
homopolymer columns (PARITY_ROADMAP.md slice 4e cluster 3).

Root cause was float-width, not FP-operation order. htslib declares its
per-base error-probability table as a C `float` array (static float
g_qual2prob[256], populated by pow(10, -i/10.) stored into a float) and
its per-query qual[] buffer is also float; the forward/backward
recurrences only promote those values to double when they enter the
double-typed HMM transition expressions. The Go port held the same table
at full float64 precision, so each entry differed from htslib's
float-rounded value by ~1e-8..1e-11. That propagated through the scaled
forward DP and flipped the integer (int)(Pr1 + .499) Phred score by one
unit at long homopolymer columns on deep reads.

Fix: round the qual2prob table through float32 so it equals htslib's
float table bit-for-bit, then promote to float64 for the arithmetic
exactly as C promotes float->double. No algorithmic behaviour changed.

Validation: a fuzz over 20k homopolymer-straddling alignments found
exactly one integer-score divergence between the float64 and float32
tables (141 vs 140); the upstream probaln.c self-test returns 140,
matching the float32 path. Pinned as TestProbalnGlocalHomopolymerFloatWidth.
The live indel parity test now diffs the full indel-AD.1 INDEL records
(I16, QS, VDB included) byte-for-byte against the upstream bcftools
binary. The calmd realn0* htslib goldens still pass.

+++ pkg/htsgo/baq/probaln.go
@@ -47,10 +47,19 @@ var ErrProbalnFailed = errors.New("baq: probaln_glocal failed")
 
 // qual2prob is the g_qual2prob table: index i holds 10^(-i/10), the error
 // probability for a Phred quality of i.
+//
+// htslib declares this table as a C `float` array (`static float
+// g_qual2prob[256]`) and computes each entry as `pow(10, -i/10.)` (a double)
+// stored into a float, so every entry is rounded to 32-bit precision before
+// use. The forward/backward recurrences then promote those floats back to
+// double for the actual arithmetic. To reproduce htslib bit-for-bit we must
+// apply the same float32 rounding here: storing full float64 values diverges
+// by ~1e-8..1e-11 per entry, which is enough to flip the single-ULP Phred
+// rounding at long homopolymer columns (the documented I16/QS indel drift).
 var qual2prob = func() [256]float64 {
 	var t [256]float64
 	for i := range t {
-		t[i] = math.Pow(10, -float64(i)/10.0)
+		t[i] = float64(float32(math.Pow(10, -float64(i)/10.0)))
 	}
 	return t
 }()
